Fix the P3 functional, and revert s/Float64/Float32/ since it breaks P3 output processing (see TODO).

+++ circus.jl
@@ -43,7 +43,7 @@ function parse_pfunctionals(args::Vector{String})
     return pfunctionals
 end
 
-function nearest_orthonormal_sinogram(input::Image{Float32})
+function nearest_orthonormal_sinogram(input::Image{Float64})
     @assert length(size(input)) == 2
 
     cols = size(input, "x")
@@ -51,7 +51,7 @@ function nearest_orthonormal_sinogram(input::Image{Float32})
 
     # Detect the offset of each column to the sinogram center
     sinogram_center = ifloor(rows / 2);
-    offset::Vector = Array(Float32, cols)
+    offset::Vector = Array(Float64, cols)
     for p in 1:cols
         median = find_weighted_median(input["y", p])
         offset[p] = median - sinogram_center
@@ -77,12 +77,12 @@ function nearest_orthonormal_sinogram(input::Image{Float32})
 end
 
 function getCircusFunction(
-    input::Image{Float32},
+    input::Image{Float64},
     pfunctional::PFunctionalWrapper)
 
     # Allocate the output matrix
     output::Vector = Array(
-        Float32,
+        Float64,
         size(input, "x")
         )
